@@ -31,7 +31,7 @@ dependencies:
   flutter:
     sdk: flutter
   geocoding: ^2.0.5
-  google_maps_flutter: ^2.0.12
+  google_maps_flutter: ^2.2.5
   flutter_map: ^3.1.0
   flutter_map_marker_cluster: any
 
@@ -54,6 +54,12 @@ dependencies:
   cloud_firestore: ^4.5.0
   permission_handler: ^10.2.0
   latlong: ^0.6.1
+  connectivity_plus: ^3.0.3
+  awesome_drawer_bar: ^1.2.0
+  # geolocator: ^9.0.2
+  google_maps_flutter_web: ^0.4.0+8
+  flutter_polyline_points: ^1.0.0
+  google_maps_routes: ^1.1.2
 
 dev_dependencies:
   flutter_test:
@@ -79,8 +85,6 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - lib/assets/images/
-    - lib/assets/images/login.png/
-    - lib/assets/images/signup.png/
 
     #   - images/a_dot_burr.jpeg
     #   - images/a_dot_ham.jpeg
