@@ -9,30 +9,32 @@ class HomePage extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-        appBar: AppBar(
-          automaticallyImplyLeading: false,
-          title: const Text('Saha-Yatri',
-              style: TextStyle(
-                fontSize: 30,
-                fontWeight: FontWeight.bold,
-                color: Colors.black,
+    return SafeArea(
+      child: Scaffold(
+          appBar: AppBar(
+            automaticallyImplyLeading: false,
+            title: const Text('Saha-Yatri',
+                style: TextStyle(
+                  fontSize: 30,
+                  fontWeight: FontWeight.bold,
+                  color: Colors.black,
+                ),
+                textAlign: TextAlign.center),
+            //make a profile button on the right side
+            actions: [
+              IconButton(
+                icon: const Icon(Icons.person),
+                onPressed: () {
+                  Navigator.pushNamed(context, '/profile');
+                },
               ),
-              textAlign: TextAlign.center),
-          //make a profile button on the right side
-          actions: [
-            IconButton(
-              icon: const Icon(Icons.person),
-              onPressed: () {
-                Navigator.pushNamed(context, '/profile');
-              },
-            ),
-          ],
-          backgroundColor: Colors.transparent,
-        ),
+            ],
+            backgroundColor: Colors.transparent,
+          ),
 
-        //homepage
+          //homepage
 
-        body: BottomBarPage());
+          body: BottomBarPage()),
+    );
   }
 }
